Add unique constraint to `last_run.command`
In b19ab98, the updating of the `last_run` table was migrated to
Postgres-native logic by using `ON CONFLICT` to perform the update
rather than an insert. This is a great change from the parent revision,
but requires a unique constraint to be on `last_run.command` so that the
update can actually occur.